[Runtime] move files into proper subdirectories
These files have so far all been in a single directory, making it harder to find files and recognize dependencies.
Additionally, the C API is now a proper layer on-top of the C++ API, exclusively used by the compiler.
