+++ src/pylir/Runtime/ExceptionHandling/ExceptionHandling.cpp
@@ -1,12 +1,11 @@
-
 //  Licensed under the Apache License v2.0 with LLVM Exceptions.
 //  See https://llvm.org/LICENSE.txt for license information.
 //  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include <iostream>
+#include <pylir/Runtime/CAPI/API.hpp>
+#include <pylir/Runtime/Objects/Objects.hpp>
 
-#include "API.hpp"
-#include "Support.hpp"
+#include <iostream>
 
 void pylir_raise(pylir::rt::PyBaseException& exception)
 {
