+++ src/pylir/Runtime/GC/GC.hpp
@@ -1,0 +1,7 @@
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#pragma once
+
+#include <pylir/Runtime/MarkAndSweep/MarkAndSweep.hpp>
