+++ src/pylir/Runtime/GC/Globals.cpp
