--- src/pylir/Runtime/GCInterface.hpp
@@ -1,9 +1,0 @@
-//  Licensed under the Apache License v2.0 with LLVM Exceptions.
-//  See https://llvm.org/LICENSE.txt for license information.
-//  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#pragma once
-
-#include <cstddef>
-
-extern "C" void* pylir_gc_alloc(std::size_t);
