+++ src/pylir/Runtime/MarkAndSweep/BestFitTree.cpp
@@ -1,10 +1,10 @@
-
 //  Licensed under the Apache License v2.0 with LLVM Exceptions.
 //  See https://llvm.org/LICENSE.txt for license information.
 //  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "BestFitTree.hpp"
 
+#include <pylir/Runtime/Objects/Objects.hpp>
 #include <pylir/Support/Util.hpp>
 
 auto pylir::rt::BestFitTree::BestFitTree::lowerBound(std::size_t size) -> std::pair<BlockHeader*, BlockHeader*>
