+++ src/pylir/Runtime/MarkAndSweep/BestFitTree.hpp
@@ -4,8 +4,7 @@
 
 #pragma once
 
-#include <pylir/Runtime/Objects.hpp>
-#include <pylir/Runtime/Pages.hpp>
+#include <pylir/Runtime/Util/Pages.hpp>
 #include <pylir/Support/Macros.hpp>
 
 #include <cstdint>
@@ -13,6 +12,8 @@
 
 namespace pylir::rt
 {
+
+class PyObject;
 
 class BestFitTree
 {
