+++ src/pylir/Runtime/MarkAndSweep/CMakeLists.txt
@@ -2,5 +2,5 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-add_library(PylirMarkAndSweep STATIC API.cpp MarkAndSweep.cpp SegregatedFreeList.cpp BestFitTree.cpp)
+add_library(PylirMarkAndSweep STATIC MarkAndSweep.cpp SegregatedFreeList.cpp BestFitTree.cpp)
 target_link_libraries(PylirMarkAndSweep PUBLIC PylirRuntime)
