+++ src/pylir/Runtime/Modules/SysModule.cpp
@@ -2,7 +2,7 @@
 //  See https://llvm.org/LICENSE.txt for license information.
 //  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "Objects.hpp"
+#include <pylir/Runtime/Objects/Objects.hpp>
 
 using namespace pylir::rt;
